refactor(game): split sprite.cpp into separate files
